Remove tests
Testing single modules like this is not practical. Some modules depend
on others, and dependencies are not managed by the modules themselves.
Also applying the modules in a clean environment is very slow. I am
removing the tests for now, until I find a better way to test the
provisioning scripts.